+++ R/plot.R
@@ -35,6 +35,10 @@ MY_THEME <- function(p, coeff = 1) {
   p + theme_bigstatsr(size.rel = coeff)
 }
 
+#' @importFrom cowplot plot_grid
+#' @export
+cowplot::plot_grid
+
 ################################################################################
 
 #' Plot method
@@ -48,10 +52,11 @@ MY_THEME <- function(p, coeff = 1) {
 #' - "loadings": plot of loadings associated with 1 Principal Component.
 #' @param nval Number of singular values to plot. Default plots all computed.
 #' @param scores Vector of indices of the two PCs to plot. Default plots the
-#' first 2 PCs.
+#' first two PCs. If providing more than two, it produces many plots.
 #' @param loadings Indices of PC loadings to plot. Default plots the
 #' first vector of loadings.
-#' @param cols If multiple vector of loadings are to be plotted, this defines
+#' @param cols Deprecated. Use `ncol` instead.
+#' @param ncol If multiple vector of loadings are to be plotted, this defines
 #' the number of columns of the resulting multiplot.
 #' @param coeff Relative size of text. Default is `1`.
 #' @param viridis Deprecated argument.
@@ -74,37 +79,67 @@ plot.big_SVD <- function(x, type = c("screeplot", "scores", "loadings"),
                          nval = length(x$d),
                          scores = c(1, 2),
                          loadings = 1,
-                         cols = 2,
+                         ncol = NULL,
                          coeff = 1,
                          viridis = TRUE,
+                         cols = 2,
                          ...) {
 
   assert_nodots()
 
   if (!missing(viridis))
     warning2("Argument 'viridis' is deprecated and will be removed.")
-
-  assert_lengths(nval, 1)
-  assert_lengths(scores, 1:2)
+  if (!missing(cols)) {
+    warning2("Argument 'cols' is deprecated and will be removed; %s",
+             "please use parameter 'ncol' instead.")
+    ncol <- cols
+  }
 
   type <- match.arg(type)
 
   if (type == "screeplot") {
 
+    assert_lengths(nval, 1)
+
     p <- MY_THEME(qplot(y = x$d[seq_len(nval)]), coeff = coeff) +
       geom_line() +
+      scale_y_log10() +
       labs(title = "Scree Plot", x = "PC Index", y = "Singular Value")
 
     `if`(nval > 12, p, p + scale_x_discrete(limits = seq_len(nval)))
 
   } else if (type == "scores") {
 
-    sc <- predict(x)
-    nx <- scores[1]
-    ny <- scores[2]
+    if (is.list(scores)) {
 
-    MY_THEME(qplot(x = sc[, nx], y = sc[, ny]), coeff = coeff) +
-      labs(title = "Scores of PCA", x = paste0("PC", nx), y = paste0("PC", ny))
+      all.p <- lapply(scores, function(scores.part) {
+        plot(x, type = "scores", scores = scores.part, coeff = coeff)
+      })
+
+      plot_grid(plotlist = all.p, ncol = ncol, scale = 0.95)
+
+    } else {
+
+      if (length(scores) > 2) {
+
+        n_plot <- floor(length(scores) / 2)
+        scores.list <- split(utils::head(scores, 2 * n_plot),
+                             rep(seq_len(n_plot), each = 2))
+        plot(x, type = "scores", scores = scores.list, ncol = ncol, coeff = coeff)
+
+      } else {
+
+        sc <- predict(x)
+        nx <- scores[1]
+        ny <- scores[2]
+
+        MY_THEME(qplot(x = sc[, nx], y = sc[, ny]), coeff = coeff) +
+          coord_fixed() +
+          labs(title = "Scores of PCA", x = paste0("PC", nx), y = paste0("PC", ny))
+
+      }
+
+    }
 
   } else if (type == "loadings") {
 
@@ -116,7 +151,7 @@ plot.big_SVD <- function(x, type = c("screeplot", "scores", "loadings"),
         p + geom_hex() + scale_fill_viridis_c()
       })
 
-      cowplot::plot_grid(plotlist = all.p, align = "hv", ncol = cols)
+      plot_grid(plotlist = all.p, align = "hv", ncol = ncol, scale = 0.95)
 
     } else {