+++ R/univLogReg.R
@@ -3,18 +3,22 @@
 univLogReg_sub <- function(X, ind, covar.train, y01.train, z0, w0,
                            ind.train, tol, maxiter) {
 
-  res <- IRLS(X, covar.train, y01.train, z0, w0,
-              ind.train, ind, tol, maxiter)
+  utils::capture.output(
+    res <- IRLS(X, covar.train, y01.train, z0, w0, ind.train, ind, tol, maxiter),
+    type = "message"
+  )
 
   # Using `glm` if not converged
-  indNoConv <- which(res$niter >= maxiter | is.na(res$estim))
+  score <- res$estim / res$std.err
+  indNoConv <- which(res$niter >= maxiter | is.na(score) | abs(score) < 1e-6)
   res$niter[indNoConv] <- NA
   for (j in indNoConv) {
     covar.train[, 1] <- X[ind.train, ind[j]]
     mod <- stats::glm(y01.train ~ covar.train - 1,
                       family = "binomial",
                       control = list(epsilon = tol, maxit = 100))
-    coeffs <- `if`(mod$converged, summary(mod)$coefficients[1, 1:2], c(NA, NA))
+    coeffs <- `if`(mod$converged && (mod$rank == ncol(covar.train)),
+                   summary(mod)$coefficients[1, 1:2], c(NA, NA))
     res$estim[j]   <- coeffs[1]
     res$std.err[j] <- coeffs[2]
   }
@@ -102,6 +106,7 @@ big_univLogReg <- function(X, y01.train,
     lpval <- stats::pnorm(xtr, lower.tail = FALSE, log.p = TRUE)
     (log(2) + lpval) / log(10)
   }
+  environment(fun.pred) <- baseenv()
 
   structure(res,
             class = c("mhtest", "data.frame"),