+++ src/utils-mat.cpp
@@ -26,37 +26,59 @@ NumericMatrix& scaling(NumericMatrix& source,
   return source;
 }
 
-/******************************************************************************/
-
-// For a squared FBM
 // [[Rcpp::export]]
-void complete2(Environment BM) {
+NumericMatrix& centering(NumericMatrix& source,
+                         const NumericVector& mean) {
 
-  XPtr<FBM> xpBM = BM["address"];
-  BMAcc<double> K(xpBM);
-  size_t m = K.ncol();
+  size_t n = source.rows();
+  size_t m = source.cols();
   size_t i, j;
 
   for (j = 0; j < m; j++)
-    for (i = j + 1; i < m; i++)
-      K(i, j) = K(j, i);
+    for (i = 0; i < n; i++)
+      source(i, j) -= mean[j];
+
+  return source;
 }
 
 /******************************************************************************/
 
-// For a square FBM
 // [[Rcpp::export]]
-void incrSup2(Environment BM,
-              const NumericMatrix& source) {
+void incr_FBM_mat(Environment BM,
+                  const NumericMatrix& mat) {
 
   XPtr<FBM> xpBM = BM["address"];
-  BMAcc<double> K(xpBM);
-  size_t m = K.ncol();
-  size_t i, j;
+  if (xpBM->matrix_type() != 8)
+    Rcpp::stop("'big_increment()' works with 'double' FBMs only.");
 
-  for (j = 0; j < m; j++)
-    for (i = 0; i <= j; i++)
-      K(i, j) += source(i, j);
+  BMAcc<double> macc(xpBM);
+
+  size_t n = macc.nrow();
+  size_t m = macc.ncol();
+
+  myassert_size(mat.rows(), n);
+  myassert_size(mat.cols(), m);
+
+  for (size_t j = 0; j < m; j++)
+    for (size_t i = 0; i < n; i++)
+      macc(i, j) += mat(i, j);
+}
+
+// [[Rcpp::export]]
+void incr_FBM_vec(Environment BM,
+                  const NumericVector& vec) {
+
+  XPtr<FBM> xpBM = BM["address"];
+  if (xpBM->matrix_type() != 8)
+    Rcpp::stop("'big_increment()' works with 'double' FBMs only.");
+
+  BMAcc<double> macc(xpBM);
+
+  size_t n = macc.size();
+  myassert_size(vec.size(), n);
+
+  for (size_t i = 0; i < n; i++)
+    macc[i] += vec[i];
 }
 
 /******************************************************************************/
@@ -64,18 +86,19 @@ void incrSup2(Environment BM,
 // For a square FBM
 // [[Rcpp::export]]
 void scaleK(Environment BM,
-                      const NumericVector& sums,
-                      const NumericVector& mu,
-                      const NumericVector& delta,
-                      int nrow) {
+            const NumericVector& sums,
+            const NumericVector& mu,
+            const NumericVector& delta,
+            int nrow) {
 
   XPtr<FBM> xpBM = BM["address"];
   BMAcc<double> K(xpBM);
-  size_t n = K.nrow();
-  size_t i, j;
 
-  for (j = 0; j < n; j++) {
-    for (i = 0; i < n; i++) {
+  size_t n = K.nrow();
+  myassert_size(K.ncol(), n);
+
+  for (size_t j = 0; j < n; j++) {
+    for (size_t i = 0; i < n; i++) {
       K(i, j) -= sums[i] * mu[j] + mu[i] * sums[j];
       K(i, j) += nrow * mu[i] * mu[j];
       K(i, j) /= delta(i) * delta(j);