+++ tests/testthat/test-SVD.R
@@ -6,7 +6,7 @@ set.seed(SEED)
 
 ################################################################################
 
-TOL <- 1e-4
+TOL <- 1e-3
 
 # function for sampling scaling
 sampleScale <- function() {
@@ -20,6 +20,19 @@ sampleScale <- function() {
 N <- 73
 M <- 43
 x <- matrix(rnorm(N * M, mean = 100, sd = 5), N)
+
+###############################################################################
+
+svd1 <- big_SVD(as_FBM(x), big_scale(), ind.row = 1:40)
+svd2 <- big_randomSVD(as_FBM(x), big_scale(), ind.row = 1:40)
+expect_equal(diffPCs(svd1$u, svd2$u), 0, tolerance = TOL)
+expect_equal(diffPCs(svd1$v, svd2$v), 0, tolerance = TOL)
+
+svd3 <- big_SVD(as_FBM(x), big_scale(), ind.row = 1:40 + 7, ind.col = 1:30 + 3)
+svd4 <- big_randomSVD(as_FBM(x), big_scale(), ind.row = 1:40 + 7,
+                      ind.col = 1:30 + 3, tol = 1e-8)
+expect_equal(diffPCs(svd3$u, svd4$u), 0, tolerance = TOL)
+expect_equal(diffPCs(svd3$v, svd4$v), 0, tolerance = TOL)
 
 ###############################################################################
 
@@ -122,3 +135,14 @@ test_that("equality with prcomp with half of half of the data", {
 })
 
 ################################################################################
+
+expect_s3_class(p <- plot(svd1, type = "scores"), "ggplot")
+expect_length(p$layers, 1)
+expect_s3_class(p <- plot(svd1, type = "scores", scores = 3:4), "ggplot")
+expect_length(p$layers, 1)
+expect_s3_class(p <- plot(svd1, type = "scores", scores = 1:8), "ggplot")
+expect_length(p$layers, 4)
+expect_s3_class(p <- plot(svd1, type = "scores", scores = 1:9), "ggplot")
+expect_length(p$layers, 4)
+
+###############################################################################